+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/_03_impute.py
@@ -1,10 +1,7 @@
 import polars as pl
 
+import projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.impute_utils as iUtils
 from polars_utils import utils
-from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
-    create_ascwds_job_role_rolling_ratio,
-    create_imputed_ascwds_job_role_counts,
-)
 
 # Set streaming chunk size for memory management - each thread (per CPU core) will load
 # in a chunk of this size.
@@ -28,12 +25,12 @@ def main(
     estimated_job_role_posts_lf = utils.scan_parquet(cleaned_data_source)
     print("Cleaned LazyFrame read in")
 
-    estimated_job_role_posts_lf = create_imputed_ascwds_job_role_counts(
+    estimated_job_role_posts_lf = iUtils.create_imputed_ascwds_job_role_counts(
         estimated_job_role_posts_lf
     )
 
-    estimated_job_role_posts_lf = create_ascwds_job_role_rolling_ratio(
-        estimated_job_role_posts_lf,
+    estimated_job_role_posts_lf = iUtils.create_ascwds_job_role_rolling_ratio(
+        estimated_job_role_posts_lf
     )
 
     utils.sink_to_parquet(
@@ -49,7 +46,10 @@ if __name__ == "__main__":
             "--cleaned_data_source",
             "Source s3 directory for merged data",
         ),
-        ("--imputed_data_destination", "Destination s3 directory for imputed data"),
+        (
+            "--imputed_data_destination",
+            "Destination s3 directory for imputed data",
+        ),
     )
     main(
         cleaned_data_source=args.cleaned_data_source,
