+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/estimate_ind_cqc_filled_posts_by_job_role.py
@@ -7,13 +7,13 @@ from typing import Final
 import polars as pl
 
 from polars_utils import utils
+from polars_utils.expressions import percentage_share
 from polars_utils.pipeline_utils import log_polars_plan, time_it
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.clean_utils import (
     nullify_job_role_count_when_source_not_ascwds,
 )
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
     ManagerialFilledPostAdjustmentExpr,
-    percentage_share,
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 from utils.column_values.categorical_column_values import PrimaryServiceType
