+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/tests/fargate/utils/test_utils.py
@@ -1,6 +1,3 @@
-from datetime import date
-from typing import Final
-
 import polars as pl
 import polars.testing as pl_testing
 import pytest
@@ -15,36 +12,6 @@ from projects._03_independent_cqc.unittest_data.polars_ind_cqc_test_file_schemas
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 PATCH_PATH = "projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils"
-
-ROW_ID: Final[str] = "id"
-EXPANDED_ID: Final[str] = "expanded_id"
-
-
-class TestGetPercentageShareRatios:
-    def test_over_groups(self):
-        expected_lf = pl.LazyFrame(
-            data=[
-                (1, "1", date(2026, 1, 1), 1, 0.3333),
-                (2, "1", date(2026, 1, 1), 2, 0.6667),
-                (3, "1", date(2026, 2, 1), 2, 0.5),
-                (4, "1", date(2026, 2, 1), 2, 0.5),
-                (5, "2", date(2026, 1, 1), 2, 0.4),
-                (6, "2", date(2026, 1, 1), 3, 0.6),
-            ],
-            schema={
-                EXPANDED_ID: pl.Int64,
-                IndCQC.location_id: pl.String,
-                IndCQC.cqc_location_import_date: pl.Date,
-                "vals": pl.Int64,
-                "ratios": pl.Float32,
-            },
-            orient="row",
-        )
-        input_lf = expected_lf.drop("ratios")
-        returned_lf = job.get_percent_share_ratios(
-            input_lf, input_col="vals", output_col="ratios"
-        ).sort(EXPANDED_ID)
-        pl_testing.assert_frame_equal(returned_lf, expected_lf, rel_tol=0.001)
 
 
 class TestPercentageShareHandlingZeroSum:
@@ -75,54 +42,6 @@ class TestPercentageShareHandlingZeroSum:
             job.percentage_share_handling_zero_sum("values").alias("pct_share")
         )
         pl_testing.assert_frame_equal(returned_lf, expected_lf)
-
-
-class TestCreateImputedASCWDSJobRoleCounts:
-    @pytest.mark.parametrize(
-        "create_imputed_ascwds_job_role_counts_data",
-        [
-            case.as_pytest_param()
-            for case in Data.create_imputed_ascwds_job_role_counts_test_cases
-        ],
-    )
-    def test_create_imputed_ascwds_job_role_counts(
-        self, create_imputed_ascwds_job_role_counts_data
-    ):
-        expected_lf = pl.LazyFrame(
-            create_imputed_ascwds_job_role_counts_data,
-            Schemas.create_imputed_ascwds_job_role_counts_expected_schema,
-            orient="row",
-        )
-        input_lf = expected_lf.drop(
-            IndCQC.ascwds_job_role_ratios,
-            IndCQC.imputed_ascwds_job_role_ratios,
-            IndCQC.imputed_ascwds_job_role_counts,
-        )
-        returned_lf = job.create_imputed_ascwds_job_role_counts(input_lf)
-        pl_testing.assert_frame_equal(returned_lf, expected_lf, rel_tol=0.0001)
-
-
-class TestCreateASCWDSJobRoleRollingRatio:
-    @pytest.mark.parametrize(
-        "create_ascwds_job_role_rolling_ratio_data",
-        [
-            case.as_pytest_param()
-            for case in Data.create_ascwds_job_role_rolling_ratio_test_cases
-        ],
-    )
-    def test_create_ascwds_job_role_rolling_ratio(
-        self, create_ascwds_job_role_rolling_ratio_data
-    ):
-        expected_lf = pl.LazyFrame(
-            create_ascwds_job_role_rolling_ratio_data,
-            Schemas.create_ascwds_job_role_rolling_ratio_expected_schema,
-            orient="row",
-        )
-        input_lf = expected_lf.drop(
-            IndCQC.ascwds_job_role_rolling_sum, IndCQC.ascwds_job_role_rolling_ratio
-        )
-        returned_lf = job.create_ascwds_job_role_rolling_ratio(input_lf)
-        pl_testing.assert_frame_equal(returned_lf, expected_lf, rel_tol=0.0001)
 
 
 class TestManagerialFilledPostAdjustmentExpr:
